Refactor: streamline MoE multi-chip example
- Keep the example focused on the end-to-end dispatch, compute, and combine path
- Remove obsolete debug docs, partial tests, and unused kernel variants
- Align README, test naming, and scratch buffer handling with the current two-chip hardware test
