--- examples/workers/l3/moe_multi_chip_experts/.gitignore
@@ -1,12 +1,0 @@
-# Log files
-*.log
-
-# Build outputs
-build_output/
-
-# Device logs
-device_log/
-
-# Analysis files
-*_analysis.md
-all_reduce.log
