--- examples/workers/l3/moe_multi_chip_experts/kernels/aiv/moe_demo_incore_0.cpp
@@ -1,108 +1,0 @@
-// Kernel Function: moe_demo_incore_0
-// Generated by PyPTO IR Compiler (PTO backend)
-
-#include <cstdint>
-
-#ifndef __gm__
-#define __gm__
-#endif
-
-#ifndef __aicore__
-#if defined(__CPU_SIM)
-#define __aicore__
-#else
-#define __aicore__ [aicore]
-#endif
-#endif
-
-#include <pto/pto-inst.hpp>
-#include "tensor.h"
-
-
-using namespace pto;
-
-
-// --- ptoas-generated code ---
-
-enum class PTOAutoSyncTailMode : int {
-  kBarrierAll = 0,
-  kSetWaitMte3ToSEvent0 = 1,
-};
-
-static __aicore__ inline void ptoas_auto_sync_tail(
-    PTOAutoSyncTailMode mode = PTOAutoSyncTailMode::kBarrierAll) {
-  switch (mode) {
-  case PTOAutoSyncTailMode::kSetWaitMte3ToSEvent0:
-    set_flag(PIPE_MTE3, PIPE_S, EVENT_ID0);
-    wait_flag(PIPE_MTE3, PIPE_S, EVENT_ID0);
-    break;
-  case PTOAutoSyncTailMode::kBarrierAll:
-  default:
-    pipe_barrier(PIPE_ALL);
-    break;
-  }
-}
-
-static __aicore__ void moe_demo_incore_0(__gm__ bfloat16_t* v1, __gm__ bfloat16_t* v2, int32_t v3, int32_t v4, int32_t v5) {
-  unsigned v6 = 0;
-  const int32_t v7 = 256;
-  const int32_t v8 = 64;
-  const int32_t v9 = 1;
-  const int32_t v10 = 16;
-  const int64_t v11 = 0;
-  using T = float;
-
-  #if defined(__DAV_VEC__)
-  set_mask_norm();
-  set_vector_mask(-1, -1);
-  Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v12 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v9, v10);
-  TASSIGN(v12, v11);
-  Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v13 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v9, v10);
-  __ubuf__ bfloat16_t* v14 = v12.data();
-  uint64_t v15 = reinterpret_cast<uint64_t>(v14);
-  TASSIGN(v13, v15);
-  pto::Shape<1, 1, 1, 1, 16> v16 = pto::Shape<1, 1, 1, 1, 16>();
-  pto::Stride<256, 256, 64, 16, 1> v17 = pto::Stride<256, 256, 64, 16, 1>();
-  GlobalTensor<bfloat16_t, pto::Shape<1, 1, 1, 1, 16>, pto::Stride<256, 256, 64, 16, 1>, pto::Layout::ND> v18 = GlobalTensor<bfloat16_t, pto::Shape<1, 1, 1, 1, 16>, pto::Stride<256, 256, 64, 16, 1>, pto::Layout::ND>(v1 + ((v6 + (unsigned) v3 * (unsigned) v7 + (unsigned) v4 * (unsigned) v8) + (unsigned) v5 * (unsigned) v10 + v6 * (unsigned) v9), v16, v17);
-  TLOAD(v13, v18);
-  set_flag(PIPE_MTE2, PIPE_MTE3, EVENT_ID0);
-  pto::Shape<1, 1, 1, 1, 16> v19 = pto::Shape<1, 1, 1, 1, 16>();
-  pto::Stride<256, 256, 64, 16, 1> v20 = pto::Stride<256, 256, 64, 16, 1>();
-  GlobalTensor<bfloat16_t, pto::Shape<1, 1, 1, 1, 16>, pto::Stride<256, 256, 64, 16, 1>, pto::Layout::ND> v21 = GlobalTensor<bfloat16_t, pto::Shape<1, 1, 1, 1, 16>, pto::Stride<256, 256, 64, 16, 1>, pto::Layout::ND>(v2 + ((v6 + (unsigned) v4 * (unsigned) v7 + (unsigned) v3 * (unsigned) v8) + (unsigned) v5 * (unsigned) v10 + v6 * (unsigned) v9), v19, v20);
-  wait_flag(PIPE_MTE2, PIPE_MTE3, EVENT_ID0);
-  TSTORE(v21, v13);
-  #endif // __DAV_VEC__
-
-  ptoas_auto_sync_tail(PTOAutoSyncTailMode::kBarrierAll);
-  return;
-}
-
-// --- Kernel entry point ---
-extern "C" __aicore__ __attribute__((always_inline)) void kernel_entry(__gm__ int64_t* args)
-{
-    // Unpack tensor: send__ssa_v0
-    __gm__ Tensor* send__ssa_v0_tensor = reinterpret_cast<__gm__ Tensor*>(args[0]);
-    __gm__ bfloat16_t* send__ssa_v0 = reinterpret_cast<__gm__ bfloat16_t*>(send__ssa_v0_tensor->buffer.addr) + send__ssa_v0_tensor->start_offset;
-
-    // Unpack tensor: recv__iter_v5
-    __gm__ Tensor* recv__iter_v5_tensor = reinterpret_cast<__gm__ Tensor*>(args[1]);
-    __gm__ bfloat16_t* recv__iter_v5 = reinterpret_cast<__gm__ bfloat16_t*>(recv__iter_v5_tensor->buffer.addr) + recv__iter_v5_tensor->start_offset;
-
-    // Unpack scalar: card_i__idx_v0
-    union { uint64_t u64; int64_t val; } card_i__idx_v0_conv;
-    card_i__idx_v0_conv.u64 = args[2];
-    int64_t card_i__idx_v0 = card_i__idx_v0_conv.val;
-
-    // Unpack scalar: expert_j__idx_v0
-    union { uint64_t u64; int64_t val; } expert_j__idx_v0_conv;
-    expert_j__idx_v0_conv.u64 = args[3];
-    int64_t expert_j__idx_v0 = expert_j__idx_v0_conv.val;
-
-    // Unpack scalar: t_idx__idx_v0
-    union { uint64_t u64; int64_t val; } t_idx__idx_v0_conv;
-    t_idx__idx_v0_conv.u64 = args[4];
-    int64_t t_idx__idx_v0 = t_idx__idx_v0_conv.val;
-
-    // Forward to ptoas-generated function
-    moe_demo_incore_0(send__ssa_v0, recv__iter_v5, card_i__idx_v0, expert_j__idx_v0, t_idx__idx_v0);
-}
