--- examples/workers/l3/moe_multi_chip_experts/kernels/aiv/moe_demo_incore_1.cpp
@@ -1,137 +1,0 @@
-// Kernel Function: moe_demo_incore_1
-// Generated by PyPTO IR Compiler (PTO backend)
-
-#include <cstdint>
-
-#ifndef __gm__
-#define __gm__
-#endif
-
-#ifndef __aicore__
-#if defined(__CPU_SIM)
-#define __aicore__
-#else
-#define __aicore__ [aicore]
-#endif
-#endif
-
-#include <pto/pto-inst.hpp>
-#include "tensor.h"
-
-
-using namespace pto;
-
-
-// --- ptoas-generated code ---
-
-enum class PTOAutoSyncTailMode : int {
-  kBarrierAll = 0,
-  kSetWaitMte3ToSEvent0 = 1,
-};
-
-static __aicore__ inline void ptoas_auto_sync_tail(
-    PTOAutoSyncTailMode mode = PTOAutoSyncTailMode::kBarrierAll) {
-  switch (mode) {
-  case PTOAutoSyncTailMode::kSetWaitMte3ToSEvent0:
-    set_flag(PIPE_MTE3, PIPE_S, EVENT_ID0);
-    wait_flag(PIPE_MTE3, PIPE_S, EVENT_ID0);
-    break;
-  case PTOAutoSyncTailMode::kBarrierAll:
-  default:
-    pipe_barrier(PIPE_ALL);
-    break;
-  }
-}
-
-static __aicore__ void moe_demo_incore_1(__gm__ bfloat16_t* v1, int32_t v2, int32_t v3, int32_t v4) {
-  RoundMode v5 = RoundMode::CAST_ROUND;
-  unsigned v6 = 0;
-  const int32_t v7 = 256;
-  const int32_t v8 = 64;
-  const float v9 = 1.0f;
-  const int32_t v10 = 1;
-  const int32_t v11 = 16;
-  const int64_t v12 = 96;
-  const int64_t v13 = 32;
-  const int64_t v14 = 0;
-  using T = float;
-
-  #if defined(__DAV_VEC__)
-  set_mask_norm();
-  set_vector_mask(-1, -1);
-  Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v15 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v10, v11);
-  TASSIGN(v15, v14);
-  Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v16 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v10, v11);
-  __ubuf__ bfloat16_t* v17 = v15.data();
-  uint64_t v18 = reinterpret_cast<uint64_t>(v17);
-  TASSIGN(v16, v18);
-  pto::Shape<1, 1, 1, 1, 16> v19 = pto::Shape<1, 1, 1, 1, 16>();
-  pto::Stride<256, 256, 64, 16, 1> v20 = pto::Stride<256, 256, 64, 16, 1>();
-  GlobalTensor<bfloat16_t, pto::Shape<1, 1, 1, 1, 16>, pto::Stride<256, 256, 64, 16, 1>, pto::Layout::ND> v21 = GlobalTensor<bfloat16_t, pto::Shape<1, 1, 1, 1, 16>, pto::Stride<256, 256, 64, 16, 1>, pto::Layout::ND>(v1 + ((v6 + (unsigned) v2 * (unsigned) v7 + (unsigned) v3 * (unsigned) v8) + (unsigned) v4 * (unsigned) v11 + v6 * (unsigned) v10), v19, v20);
-  TLOAD(v16, v21);
-  set_flag(PIPE_MTE2, PIPE_V, EVENT_ID0);
-  Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v22 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v10, v11);
-  TASSIGN(v22, v13);
-  Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v23 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v10, v11);
-  __ubuf__ float* v24 = v22.data();
-  uint64_t v25 = reinterpret_cast<uint64_t>(v24);
-  TASSIGN(v23, v25);
-  wait_flag(PIPE_MTE2, PIPE_V, EVENT_ID0);
-  TCVT(v23, v16, v5);
-  Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v26 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v10, v11);
-  TASSIGN(v26, v12);
-  Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v27 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v10, v11);
-  __ubuf__ float* v28 = v26.data();
-  uint64_t v29 = reinterpret_cast<uint64_t>(v28);
-  TASSIGN(v27, v29);
-  TEXPANDS(v27, v9);
-  Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v30 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v10, v11);
-  TASSIGN(v30, v13);
-  Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v31 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v10, v11);
-  __ubuf__ float* v32 = v30.data();
-  uint64_t v33 = reinterpret_cast<uint64_t>(v32);
-  TASSIGN(v31, v33);
-  pipe_barrier(PIPE_V);
-  TADD(v31, v23, v27);
-  Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v34 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v10, v11);
-  TASSIGN(v34, v14);
-  Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v35 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v10, v11);
-  __ubuf__ bfloat16_t* v36 = v34.data();
-  uint64_t v37 = reinterpret_cast<uint64_t>(v36);
-  TASSIGN(v35, v37);
-  pipe_barrier(PIPE_V);
-  TCVT(v35, v31, v5);
-  set_flag(PIPE_V, PIPE_MTE3, EVENT_ID0);
-  wait_flag(PIPE_V, PIPE_MTE3, EVENT_ID0);
-  TSTORE(v21, v35);
-  #endif // __DAV_VEC__
-
-  ptoas_auto_sync_tail(PTOAutoSyncTailMode::kBarrierAll);
-  return;
-}
-
-// --- Kernel entry point ---
-extern "C" __aicore__ __attribute__((always_inline)) void kernel_entry(__gm__ int64_t* args)
-{
-    // Unpack tensor: recv__iter_v12
-    __gm__ Tensor* recv__iter_v12_tensor = reinterpret_cast<__gm__ Tensor*>(args[0]);
-    __gm__ bfloat16_t* recv__iter_v12 = reinterpret_cast<__gm__ bfloat16_t*>(recv__iter_v12_tensor->buffer.addr) + recv__iter_v12_tensor->start_offset;
-
-    // Unpack scalar: expert_j__idx_v0
-    union { uint64_t u64; int64_t val; } expert_j__idx_v0_conv;
-    expert_j__idx_v0_conv.u64 = args[1];
-    int64_t expert_j__idx_v0 = expert_j__idx_v0_conv.val;
-
-    // Unpack scalar: card_i__idx_v0
-    union { uint64_t u64; int64_t val; } card_i__idx_v0_conv;
-    card_i__idx_v0_conv.u64 = args[2];
-    int64_t card_i__idx_v0 = card_i__idx_v0_conv.val;
-
-    // Unpack scalar: t_idx__idx_v0
-    union { uint64_t u64; int64_t val; } t_idx__idx_v0_conv;
-    t_idx__idx_v0_conv.u64 = args[3];
-    int64_t t_idx__idx_v0 = t_idx__idx_v0_conv.val;
-
-    // Forward to ptoas-generated function
-    moe_demo_incore_1(recv__iter_v12, expert_j__idx_v0, card_i__idx_v0, t_idx__idx_v0);
-}
