--- examples/workers/l3/moe_multi_chip_experts/kernels/aiv/moe_demo_incore_2.cpp
@@ -1,156 +1,0 @@
-// Kernel Function: moe_demo_incore_2
-// Generated by PyPTO IR Compiler (PTO backend)
-
-#include <cstdint>
-
-#ifndef __gm__
-#define __gm__
-#endif
-
-#ifndef __aicore__
-#if defined(__CPU_SIM)
-#define __aicore__
-#else
-#define __aicore__ [aicore]
-#endif
-#endif
-
-#include <pto/pto-inst.hpp>
-#include "tensor.h"
-
-
-using namespace pto;
-
-
-// --- ptoas-generated code ---
-
-enum class PTOAutoSyncTailMode : int {
-  kBarrierAll = 0,
-  kSetWaitMte3ToSEvent0 = 1,
-};
-
-static __aicore__ inline void ptoas_auto_sync_tail(
-    PTOAutoSyncTailMode mode = PTOAutoSyncTailMode::kBarrierAll) {
-  switch (mode) {
-  case PTOAutoSyncTailMode::kSetWaitMte3ToSEvent0:
-    set_flag(PIPE_MTE3, PIPE_S, EVENT_ID0);
-    wait_flag(PIPE_MTE3, PIPE_S, EVENT_ID0);
-    break;
-  case PTOAutoSyncTailMode::kBarrierAll:
-  default:
-    pipe_barrier(PIPE_ALL);
-    break;
-  }
-}
-
-static __aicore__ void moe_demo_incore_2(__gm__ bfloat16_t* v1, __gm__ bfloat16_t* v2, int32_t v3, int32_t v4) {
-  RoundMode v5 = RoundMode::CAST_ROUND;
-  unsigned v6 = 0;
-  const int32_t v7 = 256;
-  const int32_t v8 = 64;
-  const int32_t v9 = 0;
-  const float v10 = 0.0f;
-  const int32_t v11 = 1;
-  const int32_t v12 = 16;
-  const int32_t v13 = 4;
-  const int64_t v14 = 96;
-  const int64_t v15 = 64;
-  const int64_t v16 = 0;
-  using T = float;
-
-  #if defined(__DAV_VEC__)
-  set_mask_norm();
-  set_vector_mask(-1, -1);
-  Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v17 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-  TASSIGN(v17, v16);
-  Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v18 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-  __ubuf__ float* v19 = v17.data();
-  uint64_t v20 = reinterpret_cast<uint64_t>(v19);
-  TASSIGN(v18, v20);
-  set_flag(PIPE_V, PIPE_MTE2, EVENT_ID0);
-  set_flag(PIPE_V, PIPE_MTE2, EVENT_ID1);
-  TEXPANDS(v18, v10);
-  for (size_t v21 = (size_t) v9; v21 < ((size_t) v13); v21 += (size_t) v11) {
-    Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v22 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-    TASSIGN(v22, v15);
-    Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v23 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-    __ubuf__ bfloat16_t* v24 = v22.data();
-    uint64_t v25 = reinterpret_cast<uint64_t>(v24);
-    TASSIGN(v23, v25);
-    pto::Shape<1, 1, 1, 1, 16> v26 = pto::Shape<1, 1, 1, 1, 16>();
-    pto::Stride<256, 256, 64, 16, 1> v27 = pto::Stride<256, 256, 64, 16, 1>();
-    GlobalTensor<bfloat16_t, pto::Shape<1, 1, 1, 1, 16>, pto::Stride<256, 256, 64, 16, 1>, pto::Layout::ND> v28 = GlobalTensor<bfloat16_t, pto::Shape<1, 1, 1, 1, 16>, pto::Stride<256, 256, 64, 16, 1>, pto::Layout::ND>(v1 + ((v6 + (unsigned) ((int32_t) v21) * (unsigned) v7 + (unsigned) v3 * (unsigned) v8) + (unsigned) v4 * (unsigned) v12 + v6 * (unsigned) v11), v26, v27);
-    wait_flag(PIPE_V, PIPE_MTE2, EVENT_ID0);
-    TLOAD(v23, v28);
-    set_flag(PIPE_MTE2, PIPE_V, EVENT_ID0);
-    Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v29 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-    TASSIGN(v29, v14);
-    Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v30 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-    __ubuf__ float* v31 = v29.data();
-    uint64_t v32 = reinterpret_cast<uint64_t>(v31);
-    TASSIGN(v30, v32);
-    wait_flag(PIPE_MTE2, PIPE_V, EVENT_ID0);
-    pipe_barrier(PIPE_V);
-    TCVT(v30, v23, v5);
-    set_flag(PIPE_V, PIPE_MTE2, EVENT_ID0);
-    Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v33 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-    TASSIGN(v33, v16);
-    Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v34 = Tile<TileType::Vec, float, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-    __ubuf__ float* v35 = v33.data();
-    uint64_t v36 = reinterpret_cast<uint64_t>(v35);
-    TASSIGN(v34, v36);
-    pipe_barrier(PIPE_V);
-    TADD(v34, v18, v30);
-  }
-  Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v37 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-  TASSIGN(v37, v15);
-  Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v38 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-  __ubuf__ bfloat16_t* v39 = v37.data();
-  uint64_t v40 = reinterpret_cast<uint64_t>(v39);
-  TASSIGN(v38, v40);
-  pipe_barrier(PIPE_V);
-  TCVT(v38, v18, v5);
-  set_flag(PIPE_V, PIPE_MTE3, EVENT_ID0);
-  Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v41 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-  TASSIGN(v41, v15);
-  Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null> v42 = Tile<TileType::Vec, bfloat16_t, 1, 16, BLayout::RowMajor, -1, -1, SLayout::NoneBox, 512, PadValue::Null, CompactMode::Null>(v11, v12);
-  __ubuf__ bfloat16_t* v43 = v41.data();
-  uint64_t v44 = reinterpret_cast<uint64_t>(v43);
-  TASSIGN(v42, v44);
-  pto::Shape<1, 1, 1, 1, 16> v45 = pto::Shape<1, 1, 1, 1, 16>();
-  pto::Stride<64, 64, 64, 16, 1> v46 = pto::Stride<64, 64, 64, 16, 1>();
-  GlobalTensor<bfloat16_t, pto::Shape<1, 1, 1, 1, 16>, pto::Stride<64, 64, 64, 16, 1>, pto::Layout::ND> v47 = GlobalTensor<bfloat16_t, pto::Shape<1, 1, 1, 1, 16>, pto::Stride<64, 64, 64, 16, 1>, pto::Layout::ND>(v2 + ((v6 + (unsigned) v3 * (unsigned) v8) + (unsigned) v4 * (unsigned) v12 + v6 * (unsigned) v11), v45, v46);
-  wait_flag(PIPE_V, PIPE_MTE3, EVENT_ID0);
-  TSTORE(v47, v42);
-  wait_flag(PIPE_V, PIPE_MTE2, EVENT_ID0);
-  wait_flag(PIPE_V, PIPE_MTE2, EVENT_ID1);
-  #endif // __DAV_VEC__
-
-  ptoas_auto_sync_tail(PTOAutoSyncTailMode::kBarrierAll);
-  return;
-}
-
-// --- Kernel entry point ---
-extern "C" __aicore__ __attribute__((always_inline)) void kernel_entry(__gm__ int64_t* args)
-{
-    // Unpack tensor: recv__rv_v9
-    __gm__ Tensor* recv__rv_v9_tensor = reinterpret_cast<__gm__ Tensor*>(args[0]);
-    __gm__ bfloat16_t* recv__rv_v9 = reinterpret_cast<__gm__ bfloat16_t*>(recv__rv_v9_tensor->buffer.addr) + recv__rv_v9_tensor->start_offset;
-
-    // Unpack tensor: output__iter_v3
-    __gm__ Tensor* output__iter_v3_tensor = reinterpret_cast<__gm__ Tensor*>(args[1]);
-    __gm__ bfloat16_t* output__iter_v3 = reinterpret_cast<__gm__ bfloat16_t*>(output__iter_v3_tensor->buffer.addr) + output__iter_v3_tensor->start_offset;
-
-    // Unpack scalar: card_i__idx_v0
-    union { uint64_t u64; int64_t val; } card_i__idx_v0_conv;
-    card_i__idx_v0_conv.u64 = args[2];
-    int64_t card_i__idx_v0 = card_i__idx_v0_conv.val;
-
-    // Unpack scalar: t_idx__idx_v0
-    union { uint64_t u64; int64_t val; } t_idx__idx_v0_conv;
-    t_idx__idx_v0_conv.u64 = args[3];
-    int64_t t_idx__idx_v0 = t_idx__idx_v0_conv.val;
-
-    // Forward to ptoas-generated function
-    moe_demo_incore_2(recv__rv_v9, output__iter_v3, card_i__idx_v0, t_idx__idx_v0);
-}
