--- examples/workers/l3/moe_multi_chip_experts/kernels/orchestration/moe_dispatch_compute_orch.cpp
@@ -1,88 +1,0 @@
-// Orchestration Function: Dispatch + Compute (for debugging)
-//
-// This orchestration runs dispatch phase followed by compute phase.
-
-#include "runtime.h"
-#include <iostream>
-
-#include <stddef.h>
-#include <stdint.h>
-#include <stdio.h>
-
-#include "pto_orchestration_api.h"
-
-// Must match golden.py and kernel configurations
-static constexpr int64_t COUNT = 4;  // Number of tokens to process per (card, expert) pair
-static constexpr int64_t NUM_TOKENS = 10;  // Total number of tokens
-static constexpr int64_t HIDDEN_DIM = 16;  // Hidden dimension
-
-extern "C" {
-
-__attribute__((visibility("default")))
-PTO2OrchestrationConfig aicpu_orchestration_config(const ChipStorageTaskArgs& orch_args) {
-    return PTO2OrchestrationConfig{
-        .expected_arg_count = 4,  // send, recv, output, scratch (output unused)
-    };
-}
-
-__attribute__((visibility("default")))
-void aicpu_orchestration_entry(const ChipStorageTaskArgs& orch_args) {
-    // External tensors
-    Tensor ext_send = from_tensor_arg(orch_args.tensor(0));      // [num_experts][tokens][hidden]
-    Tensor ext_recv = from_tensor_arg(orch_args.tensor(1));      // [num_cards][tokens][hidden]
-    Tensor ext_output = from_tensor_arg(orch_args.tensor(2));    // [tokens][hidden] (unused)
-    Tensor ext_scratch = from_tensor_arg(orch_args.tensor(3));   // HCCL scratch buffer
-
-    // Scalar arguments
-    int64_t expert_id = static_cast<int64_t>(orch_args.scalar(0));  // Which expert this card processes
-    int64_t card_id = static_cast<int64_t>(orch_args.scalar(1));    // Which card this is
-    int64_t num_cards = static_cast<int64_t>(orch_args.scalar(2));  // Total number of cards
-    uint64_t comm_ctx_ptr = static_cast<uint64_t>(orch_args.scalar(3));  // CommContext*
-
-    printf("[Dispatch+Compute Orch] card_id=%ld expert_id=%ld num_cards=%ld\n",
-           card_id, expert_id, num_cards);
-    fflush(stdout);
-
-    PTO2_SCOPE() {
-        // === Phase 1: Dispatch ===
-        printf("[Dispatch+Compute Orch] Stage 1: Dispatch - card_id=%ld\n", card_id);
-        fflush(stdout);
-
-        Arg params_dispatch;
-        params_dispatch.add_input(ext_send);
-        params_dispatch.add_output(ext_recv);
-        params_dispatch.add_inout(ext_scratch);
-        params_dispatch.add_scalar(expert_id);
-        params_dispatch.add_scalar(num_cards);
-        params_dispatch.add_scalar(comm_ctx_ptr);
-        pto2_rt_submit_aiv_task(0, params_dispatch);  // moe_dispatch_alltoall
-
-        printf("[Dispatch+Compute Orch] Dispatch submitted for card_id=%ld\n", card_id);
-        fflush(stdout);
-
-        // === Phase 2: Compute ===
-        printf("[Dispatch+Compute Orch] Stage 2: Compute - card_id=%ld processing %d cards x %d tokens\n",
-               card_id, num_cards, COUNT);
-        fflush(stdout);
-
-        // === Phase 2: Compute ===
-        // Add 1.0 to all elements in recv[:][:4][:]
-        printf("[Dispatch+Compute Orch] Stage 2: Compute - card_id=%ld\n", card_id);
-        fflush(stdout);
-
-        Arg params_compute;
-        params_compute.add_inout(ext_recv);
-        params_compute.add_scalar(0);  // unused
-        params_compute.add_scalar(0);  // unused
-        params_compute.add_scalar(0);  // unused
-        pto2_rt_submit_aiv_task(1, params_compute);  // moe_simple_compute
-
-        printf("[Dispatch+Compute Orch] Compute submitted for card_id=%ld\n", card_id);
-        fflush(stdout);
-    }
-
-    printf("[Dispatch+Compute Orch] card_id=%ld completed\n", card_id);
-    fflush(stdout);
-}
-
-}  // extern "C"
